+++ src/console/console.h
@@ -24,8 +24,8 @@ public:
     void print();
     void update();
     void clear();
-    unsigned int get_width() const { return get_console_width(); };
-    unsigned int get_height() const { return get_console_height(); };
+    size_t get_width() const { return get_console_width(); };
+    size_t get_height() const { return get_console_height(); };
 };
 
 #endif /* SRC_CONSOLE_H_ */
